# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Actitity": {
				"enum": [
					"Checkin",
					"Checkout"
				],
				"type": "string"
			},
			"EventType": {
				"enum": [
					"Cardholder_CheckinCheckout"
				],
				"type": "string"
			},
			"CardholderCheckinoutEvent": {
				"properties": {
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"$ref": "#/components/schemas/EventType"
					},
					"message": {
						"type": "string"
					},
					"cardholderId": {
						"type": "string"
					},
					"activity": {
						"$ref": "#/components/schemas/Actitity"
					}
				},
				"required": [
					"timestamp",
					"type",
					"message",
					"cardholderId",
					"activity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EventResponse": {
				"properties": {
					"events": {
						"items": {
							"$ref": "#/components/schemas/CardholderCheckinoutEvent"
						},
						"type": "array"
					}
				},
				"required": [
					"events"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessZoneRef": {
				"description": "AccessZone represents",
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HrefMixin": {
				"properties": {
					"href": {
						"type": "string"
					}
				},
				"required": [
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PdfDefinition": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name",
					"id",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pdf": {
				"properties": {
					"href": {
						"type": "string"
					},
					"definition": {
						"$ref": "#/components/schemas/PdfDefinition"
					},
					"value": {
						"$ref": "#/components/schemas/HrefMixin"
					}
				},
				"required": [
					"href",
					"definition",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonalDataDefinition": {
				"properties": {
					"@email": {
						"$ref": "#/components/schemas/Pdf"
					},
					"@phone": {
						"$ref": "#/components/schemas/Pdf"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Status": {
				"properties": {
					"value": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CardTypeRef": {
				"description": "Reference to a card type",
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Invitation": {
				"properties": {
					"email": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"singleFactorOnly": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"mobile",
					"singleFactorOnly",
					"status",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Card": {
				"properties": {
					"href": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"cardSerialNumber": {
						"type": "string"
					},
					"issueLevel": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"type": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/CardTypeRef"
							},
							{
								"$ref": "#/components/schemas/HrefMixin"
							}
						]
					},
					"invitation": {
						"$ref": "#/components/schemas/Invitation"
					},
					"from": {
						"type": "string"
					},
					"until": {
						"type": "string"
					},
					"credentialClass": {
						"type": "string"
					},
					"trace": {
						"type": "boolean"
					},
					"lastPrintedOrEncodedTime": {
						"type": "string"
					},
					"lastPrintedOrEncodedIssueLevel": {
						"type": "number",
						"format": "double"
					},
					"pin": {
						"type": "string"
					},
					"visitorContractor": {
						"type": "boolean"
					},
					"ownedBySite": {
						"type": "boolean"
					},
					"credentialId": {
						"type": "string"
					},
					"bleFacilityId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AccessGroupRef": {
				"description": "Access Groups is what a user is assigned to to provide access to doors",
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessGroup": {
				"properties": {
					"href": {
						"type": "string"
					},
					"accessGroup": {
						"$ref": "#/components/schemas/AccessGroupRef"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"from": {
						"type": "string"
					},
					"until": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"OperatorGroupRef": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperatorGroup": {
				"properties": {
					"href": {
						"type": "string"
					},
					"operatorGroup": {
						"$ref": "#/components/schemas/OperatorGroupRef"
					}
				},
				"required": [
					"href",
					"operatorGroup"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompetencyRef": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Competency": {
				"properties": {
					"href": {
						"type": "string"
					},
					"competency": {
						"$ref": "#/components/schemas/CompetencyRef"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"expiryWarning": {
						"type": "string"
					},
					"expiry": {
						"type": "string"
					},
					"enablement": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"limitedCredit": {
						"type": "boolean"
					},
					"credit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"href",
					"competency",
					"status",
					"expiryWarning",
					"expiry",
					"enablement",
					"comment",
					"limitedCredit",
					"credit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Notifications": {
				"properties": {
					"enabled": {
						"type": "boolean"
					},
					"from": {
						"type": "string"
					},
					"until": {
						"type": "string"
					}
				},
				"required": [
					"enabled",
					"from",
					"until"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleRef": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Cardholder": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name",
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Relationship": {
				"properties": {
					"href": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleRef"
					},
					"cardholder": {
						"$ref": "#/components/schemas/Cardholder"
					}
				},
				"required": [
					"href",
					"role",
					"cardholder"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Locker": {
				"properties": {
					"href": {
						"type": "string"
					},
					"locker": {
						"$ref": "#/components/schemas/AllocatedLocker"
					},
					"from": {
						"type": "string"
					},
					"until": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"locker",
					"from",
					"until"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AllocatedLocker": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"shortName": {
						"type": "string"
					},
					"lockerBank": {
						"$ref": "#/components/schemas/Locker"
					}
				},
				"required": [
					"href",
					"name",
					"shortName",
					"lockerBank"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ElevatorGroup": {
				"properties": {
					"href": {
						"type": "string"
					},
					"elevatorGroup": {
						"$ref": "#/components/schemas/OperatorGroupRef"
					},
					"accessZone": {
						"$ref": "#/components/schemas/AccessZoneRef"
					},
					"enableCaptureFeatures": {
						"type": "boolean"
					},
					"enableCodeBlueFeatures": {
						"type": "boolean"
					},
					"enableExpressFeatures": {
						"type": "boolean"
					},
					"enableServiceFeatures": {
						"type": "boolean"
					},
					"enableService2Features": {
						"type": "boolean"
					},
					"enableService3Features": {
						"type": "boolean"
					},
					"enableVipFeatures": {
						"type": "boolean"
					}
				},
				"required": [
					"href",
					"elevatorGroup",
					"accessZone",
					"enableCaptureFeatures",
					"enableCodeBlueFeatures",
					"enableExpressFeatures",
					"enableServiceFeatures",
					"enableService2Features",
					"enableService3Features",
					"enableVipFeatures"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperatorRef": {
				"properties": {
					"href": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"href",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Redaction": {
				"properties": {
					"href": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"when": {
						"type": "string"
					},
					"before": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"redactionOperator": {
						"$ref": "#/components/schemas/OperatorRef"
					}
				},
				"required": [
					"href",
					"type",
					"when",
					"before",
					"status",
					"redactionOperator"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CardholderDetail": {
				"description": "https://gallaghersecurity.github.io/cc-rest-docs/ref/cardholders.html#definition-Cardholder-card",
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"shortName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"authorised": {
						"type": "boolean"
					},
					"lastSuccessfulAccessTime": {
						"type": "string"
					},
					"lastSuccessfulAccessZone": {
						"$ref": "#/components/schemas/AccessZoneRef"
					},
					"serverDisplayName": {
						"type": "string"
					},
					"division": {
						"$ref": "#/components/schemas/HrefMixin"
					},
					"@email": {
						"type": "string"
					},
					"@phone": {
						"type": "string"
					},
					"personalDataDefinitions": {
						"items": {
							"$ref": "#/components/schemas/PersonalDataDefinition"
						},
						"type": "array"
					},
					"cards": {
						"items": {
							"$ref": "#/components/schemas/Card"
						},
						"type": "array"
					},
					"accessGroups": {
						"items": {
							"$ref": "#/components/schemas/AccessGroup"
						},
						"type": "array"
					},
					"operatorGroups": {
						"items": {
							"$ref": "#/components/schemas/OperatorGroup"
						},
						"type": "array"
					},
					"competencies": {
						"items": {
							"$ref": "#/components/schemas/Competency"
						},
						"type": "array"
					},
					"updateLocation": {
						"$ref": "#/components/schemas/HrefMixin"
					},
					"notes": {
						"type": "string"
					},
					"notifications": {
						"$ref": "#/components/schemas/Notifications"
					},
					"relationships": {
						"items": {
							"$ref": "#/components/schemas/Relationship"
						},
						"type": "array"
					},
					"lockers": {
						"items": {
							"$ref": "#/components/schemas/Locker"
						},
						"type": "array"
					},
					"elevatorGroups": {
						"items": {
							"$ref": "#/components/schemas/ElevatorGroup"
						},
						"type": "array"
					},
					"redactions": {
						"items": {
							"$ref": "#/components/schemas/Redaction"
						},
						"type": "array"
					}
				},
				"required": [
					"authorised"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IdentityMixin": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCardholderResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/CardholderDetail"
					},
					{
						"$ref": "#/components/schemas/IdentityMixin"
					},
					{
						"$ref": "#/components/schemas/HrefMixin"
					}
				]
			},
			"CreateCardholderResponse": {
				"properties": {
					"cardholderId": {
						"type": "string"
					},
					"card": {
						"properties": {
							"expiredAt": {
								"type": "number",
								"format": "double"
							},
							"cardId": {
								"type": "string"
							}
						},
						"required": [
							"expiredAt",
							"cardId"
						],
						"type": "object"
					}
				},
				"required": [
					"cardholderId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType": {
				"enum": [
					"visitor",
					"staff",
					"vip"
				],
				"type": "string"
			},
			"AddCardToCardholderRequest": {
				"properties": {
					"cardNumber": {
						"type": "string"
					},
					"fromInMs": {
						"type": "number",
						"format": "double"
					},
					"validityPeriodInMs": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cardNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCardholderRequest": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/UserType"
					},
					"authorised": {
						"type": "boolean"
					},
					"card2Add": {
						"$ref": "#/components/schemas/AddCardToCardholderRequest"
					}
				},
				"required": [
					"userName",
					"userType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCardholderRequest": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bios-api-middleware",
		"version": "1.0.0",
		"description": "A building's inter-operational system, a.k.a, api middleware/gateway providing interfaces to help external integration with building's intranet systems.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "TBD"
		}
	},
	"paths": {
		"/acs/cardholders/{cardholderId}": {
			"get": {
				"operationId": "GetCardholder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCardholderResponse"
								}
							}
						}
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveCardholder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/acs/cardholders": {
			"post": {
				"operationId": "CreateCardholder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateCardholderResponse"
								}
							}
						}
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCardholderRequest"
							}
						}
					}
				}
			}
		},
		"/acs/cardholers/{cardholderId}": {
			"patch": {
				"operationId": "UpdateCardholder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCardholderRequest"
							}
						}
					}
				}
			}
		},
		"/acs/cardholders/{cardholderId}/activate": {
			"get": {
				"operationId": "ActivateCardholder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "authorised",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/acs/cardholders/{cardholderId}/cards": {
			"post": {
				"operationId": "AddCard2Cardholder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddCardToCardholderRequest"
							}
						}
					}
				}
			}
		},
		"/acs/cardholders/{cardholderId}/cards/{cardId}": {
			"delete": {
				"operationId": "RemoveCardFromCardholder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"ACS"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardholderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}